Speed up shopping cart. price int to double.

+++ lib/app/modules/shopping_cart/controllers/shopping_cart_controller.dart
@@ -237,7 +237,7 @@ class ShoppingCartController extends BaseController
   void getProductPrice(ShoppingCartUiModel data) {
     callDataService(
       _repository.getProductPrice(data.itemId),
-      onSuccess: (response) => data.updatePrice((response.data?.price ?? 0).toDouble()),
+      onSuccess: (response) => data.updatePrice((response.data?.price ?? 0.0)),
       onError: (error) => data.updatePrice(0.0),
       onStart: () => logger.d("Fetching product price"),
       onComplete: () => logger.d("Product price fetched"),
